models updated

+++ api/models/UnreadMessages.js
@@ -20,7 +20,7 @@ const UnreadMessages = sequlize.define('UnreadMessages', {
         type: Sequlize.INTEGER(11),
     },
     unread_status: {
-        type: Sequlize.TINYINT(2)
+        type: Sequlize.INTEGER(2)
     },
     created_at: {
         type: Sequlize.DATE,
